praja/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
supabase>=2.15.0
google-generativeai>=0.4.0
twilio>=9.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.2.0
